==> 1_Explore original data format.ipynb <==
# Finger flexion data from Kai Miller

* Data from: https://exhibits.stanford.edu/data/catalog/zk881ps0522, I believe dataset 4, 'dg.zip'
* Refer to the README.pdf in that folder for more info
* .csv data from Aashish Patel in Vikash Gilja's lab


Notes
* 1 subject of data takes a while to load and process.
* Need to notch filter line noise and apply a strong lowpass at 200Hz
* Need to epoch (maybe just compare timestamps of extrema to the cue and movement

from __future__ import print_function
%config InlineBackend.figure_format = 'retina'
%matplotlib inline

import numpy as np
import scipy as sp
import matplotlib.pyplot as plt

import seaborn as sns
sns.set_style('white')

# Load 1 subject of data
* ecog
* electrode regions
* finger flexion
* cue

# Load ecog data
from numpy import genfromtxt
subject_id = 'cc'
ecog = genfromtxt('C:/gh/data/kaidigit/csv/'+subject_id+'_ecog.csv', delimiter=',')

# Only keep first 2 channels of ecog
ecog = ecog[:2]

# Load electrode locations
elec_regions = genfromtxt('C:/gh/data/kaidigit/csv/'+subject_id+'_elecregions.csv', delimiter=',')
region_dict = {'dorsal M1':1, 'dorsal S1':3, 'ventral M1 S1':4, 'frontal':6,
               'parietal':7,'temporal':8,'occipital':9}

# Load finger data
data_finger = genfromtxt('C:/gh/data/kaidigit/csv/'+subject_id+'_dataglove.csv', delimiter=',')

finger_labels = ['thumb', 'index', 'middle', 'ring', 'pinky']

# Load cue data
data_cue = genfromtxt('C:/gh/data/kaidigit/csv/'+subject_id+'_cue.csv', delimiter=',')

cue_labels = ['rest', 'thumb', 'index', 'middle', 'ring', 'pinky']

# Define time series (Fs = 1000Hz)
Fs = 1000.
N_samples = len(ecog[0])
t_max = N_samples/float(Fs)
t = np.arange(0,t_max,1/float(Fs))

elec_regions

# Plot time series

e1 = 0
t_lim = (5,7)
t_plt1 = t[np.logical_and(t>=t_lim[0], t<t_lim[1])]
x_plt1 = ecog[0][np.logical_and(t>=t_lim[0], t<t_lim[1])]
t_plt2 = t[np.logical_and(t>=t_lim[0], t<t_lim[1])]
x_plt2 = ecog[1][np.logical_and(t>=t_lim[0], t<t_lim[1])]

plt.figure(figsize=(8,2))
plt.plot(t_plt1,x_plt1*.0298,'k')
plt.plot(t_plt2,x_plt2*.0298,'b')
plt.xlabel('Time (s)')
plt.ylabel('Voltage (uV)')

# Plot power spectrum

# Calculate PSD
from misshapen import nonshape
f, psd = nonshape.f_psd(ecog[e1], Fs, 'welch',welch_params={'nperseg': 10000, 'window': 'hanning', 'noverlap': None})
f2, psd2 = nonshape.f_psd(ecog[1], Fs, 'welch',welch_params={'nperseg': 10000, 'window': 'hanning', 'noverlap': None})

# Plot psd
flim = (0,10)
plt.figure(figsize=(8,4))
plt.semilogy(f[np.logical_and(f>=flim[0],f<flim[1])],psd[np.logical_and(f>=flim[0],f<flim[1])],'k')

# Calculate PSD
from misshapen import nonshape
x_psd = ecog[1]
f, psd = nonshape.f_psd(x_psd, Fs, 'welch',welch_params={'nperseg': 10000, 'window': 'hanning', 'noverlap': None})
# Plot psd
flim = (0,10)
plt.figure(figsize=(8,4))
plt.semilogy(f[np.logical_and(f>=flim[0],f<flim[1])],psd[np.logical_and(f>=flim[0],f<flim[1])],'k')

elec_regions

# Plot finger and cue data

t_lim = (30,40)
t_plt = t[np.logical_and(t>=t_lim[0], t<t_lim[1])]
cue_plt = data_cue[np.logical_and(t>=t_lim[0], t<t_lim[1])]
fingers_plt = data_finger[:,np.logical_and(t>=t_lim[0], t<t_lim[1])]

plt.figure(figsize=(8,4))
plt.subplot(2,1,1)
plt.plot(t_plt,cue_plt,'k')
plt.ylim((-.5,5.5))
plt.ylabel('Cue')
plt.subplot(2,1,2)
plt.plot(t_plt,fingers_plt[0],'k',label=finger_labels[0])
plt.plot(t_plt,fingers_plt[1],'b',label=finger_labels[1])
plt.plot(t_plt,fingers_plt[2],'r',label=finger_labels[2])
plt.plot(t_plt,fingers_plt[3],'g',label=finger_labels[3])
plt.plot(t_plt,fingers_plt[4],'y',label=finger_labels[4])
plt.xlabel('Time (s)')
plt.ylabel('Finger flexion')
plt.legend(loc='best')
