==> 2_Separate data.ipynb <==
# Separate data by electrode and finger
Data format should be:
* data2/dg/subject
    * region/electrode/time
    * cue
    * fingers

from __future__ import print_function
import numpy as np
import os
from numpy import genfromtxt

# Define variables to loop through
all_subjects = ['bp','cc','zt','jp','ht','mv','wc','wm','jc']
all_elec_regions = np.arange(10)
all_fingers = np.arange(5)

# Define start and end directories
dir_start = 'C:/gh/data/kaidigit/csv/'
dir_end = 'E:/data2/dg/'

# Loop through subjects
for s in all_subjects:
    print('subject '+s)
    
    # Make a subject directory if it does not yet exist
    if not os.path.exists(dir_end+s):
        os.makedirs(dir_end+s)
        
    # resave cue data
    data_cue = genfromtxt(dir_start+s+'_cue.csv', delimiter=',')
    np.save(dir_end+s+'/cue.npy', data_cue)
    print('cue')
    
    # Save finger data
    data_finger = genfromtxt(dir_start+s+'_dataglove.csv', delimiter=',')
    for i in range(np.shape(data_finger)[0]):
        np.save(dir_end+s+'/finger'+str(i)+'.npy', data_finger[i])
    print('fingers')
        
    # Load ecog and location data
    data_regions = genfromtxt(dir_start+s+'_elecregions.csv', delimiter=',')
    data_ecog = genfromtxt(dir_start+s+'_ecog.csv', delimiter=',')
    print('load ecog')
    
    # For each region, save that ecog channel
    for i in range(len(data_regions)):
        # Determine channel location
        chan_loc = int(data_regions[i])
        
        # Create a folder for that channel location if does not already exist
        if not os.path.exists(dir_end+s+'/'+str(chan_loc)):
            os.makedirs(dir_end+s+'/'+str(chan_loc))
            
        # Save the channel
        np.save(dir_end+s+'/'+str(chan_loc)+'/'+str(i)+'.npy', data_ecog[i])
        
    print('save ecog')
