==> 6_Save cue plot.ipynb <==
# Save plots for cues - all subjects

from __future__ import print_function
%config InlineBackend.figure_format = 'retina'
%matplotlib inline

import numpy as np
import scipy as sp
import matplotlib.pyplot as plt
import os

import glob
# Define data of interest
data_dir = 'E:/data2/dg/'
image_dir = data_dir+'/cue_plot'
subj_name = 2
subjects = glob.glob(data_dir+'/*')
all_subjects = []

if not os.path.exists(data_dir+"/cue_plot"):
    os.makedirs(data_dir+"/cue_plot")

for sub in subjects:
    if len(sub) == len(data_dir)+subj_name:
        file_name = sub[-subj_name:]
        all_subjects.append(file_name)
        


for i in all_subjects:
    cue = np.load('E:/data2/dg/'+i+'/cue.npy')
    finger0 = np.load('E:/data2/dg/'+i+'/finger0.npy')
    finger1 = np.load('E:/data2/dg/'+i+'/finger1.npy')
    finger2 = np.load('E:/data2/dg/'+i+'/finger2.npy')
    finger3 = np.load('E:/data2/dg/'+i+'/finger3.npy')
    finger4 = np.load('E:/data2/dg/'+i+'/finger4.npy')

    plt.figure(figsize=(8,4))
    plt.subplot(2,1,1)
    plt.plot(cue[:70000])
    plt.ylim((-.5,5.5))
    plt.subplot(2,1,2)
    plt.plot(finger0[:70000])
    plt.plot(finger1[:70000])
    plt.plot(finger2[:70000])
    plt.plot(finger3[:70000])
    plt.plot(finger4[:70000])
    plt.savefig(image_dir+ "/"+i+"_cue.png")
    plt.close()
