==> 6_Save PSD plot.ipynb <==
#Load data, calculate PSD and save the plots

from __future__ import print_function
%config InlineBackend.figure_format = 'retina'
%matplotlib inline

import numpy as np
import scipy as sp
import matplotlib.pyplot as plt

import seaborn as sns
sns.set_style('white')
import sys
sys.path.append("C:/Users/user/Desktop/ipynb/misshapen/")
sys.path.append("C:/Users/user/Desktop/ipynb/")
import misshapen
import nonshape
import shape

import os
from numpy import genfromtxt


import glob
# Define data of interest
data_dir = 'E:/data2/dg/'
image_dir = data_dir+'/psd_plot'
subj_name = 2
subjects = glob.glob(data_dir+'/*')
all_subjects = []

for sub in subjects:
    if len(sub) == len(data_dir)+subj_name:
        file_name = sub[-subj_name:]
        all_subjects.append(file_name)


Fs = 1000
region = [1,3,4,6,7,8,9]

if not os.path.exists(data_dir+"/psd_plot"):
    os.makedirs(data_dir+"/psd_plot")
# Determine file name


# Loop through all subj
for subj in all_subjects:
    for r in region:
        if os.path.exists(data_dir+subj+'/'+str(r)+'/'):
            ecog_files = glob.glob(data_dir+subj+'/'+str(r)+'/*')
            files_num = np.arange(len(ecog_files))
            if len(files_num) <= 8:
                fig = plt.figure(figsize=(16,8))
            else:
                fig = plt.figure(figsize=(16,len(files_num)))

            for i in files_num:
                if len(ecog_files)<3:
                    subF = fig.add_subplot(len(files_num), 1, (i+1))
                else:
                    subF = fig.add_subplot(len(files_num)/2+1, 2, (i+1))
                # Load data
                x = np.load(ecog_files[i])
                f, psd = nonshape.f_psd(x, Fs, 'welch',welch_params={'nperseg': 10000, 'window': 'hanning', 'noverlap': None})
                # Plot psd
                flim = (0,100)
        
                subF.semilogy(f[np.logical_and(f>=flim[0],f<flim[1])],psd[np.logical_and(f>=flim[0],f<flim[1])],'k')
                subF.set_xlabel('frequency (hz)')
                subF.set_ylabel('power (uV^2/Hz)')
        
            fig.tight_layout()
            if r == 1:
                fig.suptitle('dorsal M1', fontsize = 20)
            elif r == 3:
                fig.suptitle('dorsal S1', fontsize = 20)
            elif r == 4:
                fig.suptitle('ventral sensorimotor (M1+S1)', fontsize = 20)
            elif r == 6:
                fig.suptitle('frontal (non-rolandic)', fontsize = 20)
            elif r == 7:
                fig.suptitle('parietal (non-rolandic)', fontsize = 20)
            elif r == 8:
                fig.suptitle('temporal', fontsize = 20)
            elif r == 9:
                fig.suptitle('occipital', fontsize = 20)
            fig.subplots_adjust(top=0.9)
            fig.savefig(image_dir+ "/"+subj+"_"+str(r)+".png")
            plt.close()
