==> 10_Average for each trial.ipynb <==
%config InlineBackend.figure_format = 'retina'
%matplotlib inline

import numpy as np
import scipy as sp
import pandas as pd
import matplotlib.pyplot as plt

import seaborn as sns
sns.set_style('white')

import sys
#sys.path.append("C:/Users/user/Desktop/ipynb/misshapen/")
sys.path.append("C:/Users/Yimeng/Documents/GitHub/")
from neurodsp import shape

import util
import imp
imp.reload(util)
from __future__ import print_function

import os
import glob


# Load behavioral dataframe

dfmove = pd.read_csv('behavior_all.csv',index_col=0)
dfmove

# Load neural data

Fs = 1000 # sampling rate
f_range = (10,30) # frequency range of oscillation of interest

dfs=[]
data_paths, all_regions = util.load_paths()

for key in data_paths:
    df_subj = dfmove[dfmove['subject']==key]
    for i in data_paths[key]['1']:
        electrode=i.split('\\')[1]
        electrode = electrode[:-4]
        x = np.load(i) # voltage series
        
        df = shape.features_by_cycle(x, Fs, f_range, center_extrema='T' )
        
        df['subject']=key

        df_no_bad_trials = util.drop_bad_trials(df_subj,1500,'delay_start',True,True)
        df_no_bad_trials = util.drop_bad_trials(df_subj,2000,'delay_end',True,True)
        #rdsym,average,sem=util.plot_shape(x,df,df_no_bad_trials,-1000,1000,Fs,key,electrode)
        dfs.append(df)

subj_keys=list(data_paths.keys())

dfn=pd.concat(dfs) 
dfn.to_csv('./neural_all_neurodsp.csv')

dfneural = pd.read_csv('neural_all_neurodsp.csv',index_col=0) 
dfneural

def calculate_average(trial,label):
    total = 0
    if(len(trial)!=0):
        for num in trial[label]:
            total += num;
        return total/len(trial)
    return 0

dfneural_avg = dfneural
dfneural_nomove=pd.DataFrame()
dfneural_nomove

interval = 500
for subj in data_paths:
    df_subj_move = dfmove[dfmove['subject']==subj]
    df_subj_neural=dfneural[dfneural['subject']==subj]
    start=df_subj_move['move_start']
    end = df_subj_move['move_end']

    for index in range(len(start)):
        average_nomove(index,start,end,df_subj_neural,dfneural_nomove)
        average_move(index,start,end,df_subj_neural,dfneural_avg)
        


dfneural_avg
dfneural_avg.to_csv('./dfneural_avg.csv')

dfneural_nomove
dfneural_nomove.to_csv('./dfneural_nomove.csv')

def average_nomove(index,start,end,df_subj_neural,dfneural_nomove):
        try:
            if index != len(start)-1:
                middle = (end[index]+start[index+1])/2
                nomove= df_subj_neural[(df_subj_neural['sample']>=middle-interval) & (df_subj_neural['sample']<=middle+interval)]
            
                avg_nomove=calculate_average(nomove,'rdsym_time')
                dfneural_nomove=pd.concat([dfneural_nomove,nomove],ignore_index=True)
        
                for sample in nomove['sample']:
                    dfneural_nomove.loc[dfneural_nomove['sample']==sample,'average rdsym']=avg_nomove
                
        except KeyError:
            avg=0  
        

def average_move(index,start,end,df_subj_neural,dfneural_avg):
        try:
            trial = df_subj_neural[(df_subj_neural['sample']<=end[index]) & (df_subj_neural['sample']>=start[index])]
        
            total = 0
            avg = 0
            if(len(trial)!=0):
                avg = calculate_average(trial,'rdsym_time')
            
                for sample in trial['sample']:
                    dfneural_avg.loc[dfneural_avg['sample']==sample,'average rdsym']=avg
        except KeyError:
            avg=0
