==> 11_Logistic Regression _all subject.ipynb <==
# Train logistic regression modifier on all subjects

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import util

from sklearn import metrics
from sklearn import linear_model, datasets
from sklearn.model_selection import cross_val_score

# Useful function

def load_move_nomove_each(move_csv, nomove_csv):
    """ Load the movement dataframe and nomovement dataframe and label them
    
        Parameter:
            move_csv (String): The name of the move csv
            nomove_csv (String): The name of the nomove csv
            
        Return:
            df (pandas dataframe): The dataframe that has both data labelled
    """
    # Load feature dataframes 
    dfmove = pd.read_csv(move_csv,index_col=0)
    dfnomove = pd.read_csv(nomove_csv,index_col=0)

    # Remove non-feature cols
    dfmove.drop(['end_time', 'start_time'], axis=1, inplace=True)
    dfnomove.drop(['end_time', 'start_time'], axis=1, inplace=True)
    
    # Set label for each dataframe
    dfmove['move']=1
    dfnomove['move']=0

    # concatenate dfs
    df = pd.concat([dfmove, dfnomove])
    df.reset_index(inplace=True, drop=True)
    df.dropna(inplace=True,axis=0)
    return df

def calculate_test_train_each(df, fold, subj):
    """ Calculate the n-fold cross validation and also calculate the accuracy 
        on the training set. Print the confusion matrix
        
        Parameter:
            df (pandas dataframe): The data frame for both move and nomove
            
            fold: The number that we want to calculate for cross validation
            
            subj: The subject we are calculating for 
            
        Return:
            A dictioanry containing the results
            A confusion matrix
    """
    
    result={}
    # get the X and Y to feed log reg
    X = df[list(df.keys())[:-1]].values
    Y = df['move'].values
    
    # Caluculate cross validation
    logreg = linear_model.LogisticRegression(C=1e5)
    scores = cross_val_score(logreg, X, Y, cv=fold)
    result['cross_val'] = np.mean(scores) # Average accuracy
    
    # Calculate Confusion matrix and accuracy on training set
    logreg.fit(X,Y)
    y_pred = logreg.predict(X)
    
    result['training']=np.mean(Y==y_pred)
    return result, metrics.confusion_matrix(Y, y_pred)

# Get all subjects
all_subjects = util.get_subjects_all('C:/data2/dg')

result_all={}
for subj in all_subjects:
    move = 'neural_move_'+subj+'.csv'
    nomove = 'neural_nomove_'+subj+'.csv'
    
    df = load_move_nomove_each(move,nomove)
    result_each,metric=calculate_cross_val(df,5,subj)
    print(subj)
    print(metric)
    result_all[subj]=result_each
    

result_all
