==> 11_Logistics regression.ipynb <==
import numpy as np
import matplotlib.pyplot as plt
from sklearn import linear_model, datasets
from sklearn.model_selection import cross_val_score



%config InlineBackend.figure_format = 'retina'
%matplotlib inline

import numpy as np
import scipy as sp
import pandas as pd
import matplotlib.pyplot as plt

import seaborn as sns
sns.set_style('white')

import sys
#sys.path.append("C:/Users/user/Desktop/ipynb/misshapen/")
sys.path.append("C:/Users/Yimeng/Documents/GitHub/")
from misshapen import shape, nonshape

import util
import imp
imp.reload(util)
from __future__ import print_function

import os
import glob

# Load dataframe

df = pd.read_csv('dfneural_avg.csv',index_col=0)
df_nomove = pd.read_csv('dfneural_nomove.csv',index_col=0)

# import some data to play with
iris = datasets.load_iris()
X = iris.data[:, :2]  # we only take the first two features.
Y = iris.target

h = .02  # step size in the mesh

logreg = linear_model.LogisticRegression(C=1e5)

scores = cross_val_score(logreg, X, Y, cv=5)
np.mean(scores) # Average accuracy
